requirements.txt
requests
python-dateutil
celery
redis
flower
python-dotenv
click